Allow rules to be callables, unify calling convention, pass data
The rules for a given dqt can just be a callable, in which case it is
called and its result is the result of the rules.

All user functions are now called with the same set of arguments.

You can pass data in, in the form of extra keyword arguments.