Everything is a function: infrastructure & commentary
Infrastructure: there is now no special recursion in the walker, but
rather a mechanism for calling functions with arguments, and a
function which recursively invokes the walker.

Commentary: at least some description of what the walker builds, and
some notes on whether it would be worth abstracting over all the
dq.inx.* stuff, which I've decided not to do yet.