Support for multiple instances of a CMORvar with the same name
This happens, it turns out, and previously the walker would just
silently clobber subseqent instances.  Now it builds a tuple of walks,
so the structure looks like

  {miptable: {varname: (walk, ...), ...}, ...}

rather than the old

  {miptable: {varname: walk, ...}, ...}

twalk needed some changes to deal with this (it assumed that the only
places lists or tuples occured they would be collections of strings).

Some small changes to twalk and jwalk to stop them puking a backtrace
on errors.

Also modify dqcsv to support the new walk_dq result structure (tuples)