djq: toplevel
djq.toplevel implements the toplevel for the script: the code which
deals with the reading the request and replying to it, and is the
place where exceptions should be handled.

Some tests for it as well (currently only failure tests).

A stub for djq.variables which will compute the variables (ie the guts
of it).