djq: support for all/no experiments
variables.py reworked for this (still not actually likely to be
correct), and some new test samples.

Also some changes to exceptions: as well as Scram (which now isn't
used, but could be for debugging) there is now Disaster which is a
generic 'I give up' internal error which anything can raise and which
should normally induce a catastrophe.