djq: all mutable global state is now thread-local or otherwise safe
This means that a python-callable interface should be thread-safe.

The cache of loaded DREQs in djq.toplevel is still global, but has
been changed to use a two-level cache indexed by root and tag.
Assuming dicts are low-level safe (which they are) this is safe (it
may do extra harmless work because of a race) but allows loaded DREQs
to be shared between threads which share roots.

Other mutable global state -- the default root and tag in djq.load and
the noisiness flags in djq.load.noise -- are now completely
thread-local.