diagnostic reviewers: pseudorandomly pick victims
Where there are several possible victims (several people who have
reviewed no more than the smallest number of diagnostics so far) this
now picks one using a PRNG.  You can seed this with -s to make the
results, I hope, deterministic, while not, for instance, picking
people alphabetically first all the time.

Also noted that this is not in fact the best assignment in all cases
as it only depends on the assignments it has already seen (see
commentary).