Some extra debugging output
Report information about dqroot & dqtag in slightly too much detail

Report the path for the djq package & later the dreqPy package, which
helps check that the versions loaded are the ones that should be
loaded.

Print the XML & config paths jus before calling the loader.

Finally report the version of the dreq actually loaded.