Suppress the manifest when loading the dreq
It looks as though, in recent versions of dreqPy (how recent?), if
you don't explicitly set the manifest to None, then it will use it
in preference to any arguments you give it, thus causing it to load
a version of the dreq of its choosing rather than one you asked for.

It would be nicer to compute the manifest path and just provide that.
But the manifest did not exist at all before 01.beta.29 and it wasn't
correct until 01.beta.30, so doing this would prevent loading those old
versions, which would be bad.

Resolves #3