Add metadata to the reply
There is a new field called reply-metadata which has a lot of useful
metadata about the reply, none of which is specified.  This corresponds
to a lot of the debug output which recently arrived, and that has now
gone.

Implementationally this uses a new fluid so various parts of the thing
can add information to it.  This is in a new module, metadata.py, which
is not exported at all from the package. I am still not really happy
with constructing coherent namespaces in Python.

Resolves #4