Support for direct paths
Rather than providing a root and tag, you can specify a path where the
XML files are assumed to live, and it will look there, ignoring root
& tag.  This works for djq, all-requests and the Python API.

djq and all-requests have a new -p option which specifies the path.
The Python API has dqpath and default_dqpath which do the same thing.
If you specify a path then the root and tag are completely ignored.

There may still be problems with things which appear in the response
(the dreq version, which can come from the tag in some way).  the root
and tag are still reported in the response: you need to know that, if
dqpath is there then they mean nothing.

The documentation should be up to date

This is part of #5
(It may be all of it.)