Update the xslx sanity checker
It now can process newer spreadsheets, but it *won't* process old ones
any more as they have changed incompatibly

It also supports direct paths at the module level (this is somewhat
hacky, as it needs to guess where the spreadsheet is).  The scripty
bit *doesn't* support direct paths.

The sanity tests still fail, but I am not sure that is djq's fault any
more.

This is part of #2 and #5