Fix a stupidity with FH closing
There were at least two bugs:
- it tried to refer to stderr when stderr had not been imported (this
  must have been some previous change);
- it tried to flush stderr & stdout without checking they were open.

These were all masked because it never reported anything since this was
all last-gasp code: it just died with EX_IOERR (which it should have
done).  It now checks things properly (and imports the names), and
also if debug_level() is 2 or more it will reraise the exception.