Add memoization support
The djq.low.memoize module can memoize functions, using a fluid to
hold the memos.  By default it memoizes single-argument functions, and
the argument needs to be hashable.  There is an optional 'key'
function which can extract a hashable key from an unhashable argument,
and additionally if the 'spread' option is given the wrapper gets a
tuple of the arguments to the wrapped function which it then spreads
to call the wrapped function if need be: tuples are hashable if all
their elements are, so it can memoize some multiple-argument
functions.

There are tests which are quite slow to run (because some of them are
trying to test functions which would not be computable in reasonable
time without memoization) and which also may not be thread-safe since
they use SIGALRM to time-limit computations.

The appropriate fluid is bound in the toplevel call, but nothing in
djq is actually memoized yet.

I had hoped that this would allow big improvements to performance.  It
still may but they are not as easy to find as I thought.