Bind fluids in a better way
djq and cci now pass arguments to their process_* functions, which
then bind the appropriate fluids.  Previously they set global values
for all the fluids (although they did not know they were doing this
really) and then the process_* functions rebound them, which was ugly.

There are two exceptions: the verbosity and debug fluids are still
assigned at the program level, as they need to be right very early.

This is a pretty major structural change: the programs are now better
I think, but there may well be lurking horrors.